Revisi dari email 13 feb
- perubahan font
- perubahan header menu dan footer
- penambahan + di wedding data client
- perubahan about us

+++ app/Http/Controllers/CoupleController.php
@@ -262,7 +262,9 @@ class CoupleController extends Controller
             $filename = $pdf->getClientOriginalName();
             $destinationPath = public_path('/images/upload/pdf');
 
-            unlink($destinationPath.'/'.$old_pdf);
+            if ($old_pdf) {
+                unlink($destinationPath.'/'.$old_pdf);
+            }
             
             $proses = $request->file('pdf')->move($destinationPath, $filename);
             $mr = $filename;
